Fixed Code on Kirill's recommendation.

+++ Assets/Zombieland/GameScene0/CharacterModule/BuffDebuffModule/Scripts/BuffDebuffData.cs
@@ -7,8 +7,8 @@ namespace Zombieland.GameScene0.CharacterModule.BuffDebuffModule
         public string IconID { get; private set; } // Serializable
         public string PreabID { get; private set; } // Serializable
         public VFXPosition VFXPosition { get; private set;} // Serializable
-        public int LifeTime { get; private set; } // Serializable
-        public int Interval { get; private set; } // Serializable
+        public float LifeTime { get; private set; } // Serializable
+        public float Interval { get; private set; } // Serializable
         public DirectImpactData DirectImpactData { get; set; }
         public ICharacterController ImpactTarget { get; set; }
         public ICharacterController Owner { get; set; }
